% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply.R
\name{image_apply}
\alias{image_apply}
\title{Apply a Function over Image Frames}
\usage{
image_apply(image, FUN, ...)
}
\arguments{
\item{image}{object returned by \code{image_read}}

\item{FUN}{a function to be called on each frame in the image}

\item{...}{additional parameters for \code{FUN}}
}
\description{
\code{image_apply} calls a transformation function to each frame of the image and
joints the result back into a single image. Because most operations are already
vectorized this is often not needed. Note that \code{FUN} should return an image.
To apply other kinds of functions simply use \link{lapply} or \link{vapply}, etc.
}
