% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EBImage.R
\name{as_EBImage}
\alias{as_EBImage}
\title{Convert to EBImage}
\usage{
as_EBImage(image)
}
\arguments{
\item{image}{magick image object returned by \code{\link[=image_read]{image_read()}} or \code{\link[=image_graph]{image_graph()}}}
}
\description{
Convert a Magck image to \href{https://bioconductor.org/packages/release/bioc/html/EBImage.html}{EBImage}
class. Note that EBImage only supports multi-frame images in greyscale.
}
