% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{image_attr_antialias}
\alias{image_attr_animationdelay}
\alias{image_attr_antialias}
\alias{image_attr_backgroundcolor}
\alias{image_attr_boxcolor}
\alias{image_attr_density}
\alias{image_attr_fillcolor}
\alias{image_attr_font}
\alias{image_attr_fontsize}
\alias{image_attr_format}
\alias{image_attr_gifmethod}
\alias{image_attr_label}
\alias{image_attr_matte}
\alias{image_attr_mattecolor}
\alias{image_attr_pencolor}
\alias{image_attr_quality}
\alias{image_attr_quantize}
\title{Image Attributes}
\usage{
image_attr_antialias(image, antialias = NULL)

image_attr_animationdelay(image, delay = NULL)

image_attr_backgroundcolor(image, color = NULL)

image_attr_boxcolor(image, color = NULL)

image_attr_density(image, geometry = NULL)

image_attr_fillcolor(image, color = NULL)

image_attr_font(image, font = NULL)

image_attr_fontsize(image, fontsize = NULL)

image_attr_gifmethod(image, method = NULL)

image_attr_label(image, label = NULL)

image_attr_format(image, format = NULL)

image_attr_matte(image, matte = NULL)

image_attr_mattecolor(image, color = NULL)

image_attr_pencolor(image, color = NULL)

image_attr_quality(image, quality = NULL)

image_attr_quantize(image, quantize = NULL)
}
\arguments{
\item{image}{object returned by \code{image_read}}

\item{antialias}{Control antialiasing of TrueType fonts (TRUE/FALSE)}

\item{delay}{time in 1/100ths of a second (0 to 65535) which must expire before
displaying the next image in an animated sequence.}

\item{color}{a valid \href{https://www.imagemagick.org/Magick++/Color.html}{color string}}

\item{geometry}{Vertical and horizontal resolution for reading PDF. Defaults is \code{"72x72"}}

\item{font}{Text rendering font. To use a TrueType font, precede the TrueType filename with an @.}

\item{fontsize}{Text rendering font point size}

\item{method}{one of: 0 (unspecified) 1 (do not dispose), 3 (overwrite with background color)
or 4 (overwrite with previous graphic.}

\item{label}{string with image label}

\item{format}{string with image format such as \code{png}, \code{gif}, etc.}

\item{matte}{if the image has transparency (TRUE/FALSE).}

\item{quality}{JPEG/MIFF/PNG compression level (default 75)}

\item{quantize}{Preferred number of colors in the image. The actual number of colors '
in the image may be less than your request, but never more.}
}
\description{
Functions to get or set Image attributes. Attributes typically do not
alter the image itself, but can affect how images get rendered or transformed.
Each function modifies the image in place and returns a vector with current
attribute values.
}

