% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoview.R
\name{autoviewer}
\alias{autoviewer}
\alias{autoviewer_enable}
\alias{autoviewer_disable}
\title{RStudio AutoViewer}
\usage{
autoviewer_enable()

autoviewer_disable()
}
\description{
This enables a \link{addTaskCallback} that automatically updates the viewer after
the state of a magick graphics device has changed. This is enabled by default in
RStudio.
}
\examples{
# Only has effect in RStudio (or other GUI with a viewer):
autoviewer_enable()

img <- magick::image_graph()
plot(1)
abline(0, 1, col = "blue", lwd = 2, lty = "solid")
abline(0.1, 1, col = "red", lwd = 3, lty = "dotted")

autoviewer_disable()
abline(0.2, 1, col = "green", lwd = 4, lty = "twodash")
abline(0.3, 1, col = "black", lwd = 5, lty = "dotdash")

autoviewer_enable()
abline(0.4, 1, col = "purple", lwd = 6, lty = "dashed")
abline(0.5, 1, col = "yellow", lwd = 7, lty = "longdash")
}
