% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coder.R
\name{coder_info}
\alias{coder_info}
\title{Coder Info}
\usage{
coder_info(format)
}
\arguments{
\item{format}{image format such as \code{png}, \code{tiff} or \code{pdf}.}
}
\description{
Provide information regarding ImageMagick support for an image format.
Raises an error for unsupported formats.
}
\examples{
coder_info("png")
coder_info("jpg")
coder_info("pdf")
coder_info("tiff")
coder_info("gif")
}
\references{
\url{https://www.imagemagick.org/Magick++/CoderInfo.html}
}

