% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edges.R
\name{image_edge}
\alias{image_edge}
\alias{image_canny}
\alias{image_hough_draw}
\alias{image_hough_txt}
\title{Edge / Line Detection}
\usage{
image_edge(image, radius = 1)

image_canny(image, geom = "0x1+10\%+30\%")

image_hough_draw(image, geom = "5x5+20", color = "red",
  bg = "transparent", size = 3, overlay = FALSE)

image_hough_txt(image, geom = "5x5+20", format = c("mvg", "svg"))
}
\arguments{
\item{image}{magick image object returned by \code{\link[=image_read]{image_read()}} or \code{\link[=image_graph]{image_graph()}}}

\item{radius}{edge size in pixels}

\item{geom}{geometry string, see details.}

\item{color}{a valid \href{https://www.imagemagick.org/Magick++/Color.html}{color string} such as
\code{"navyblue"} or \code{"#000080"}}

\item{bg}{background color}

\item{size}{size in points to draw the line}

\item{overlay}{composite the drawing atop the input image. Only for \code{bg = 'transparent'}.}

\item{format}{output format of the text, either \code{svg} or \code{mvg}}
}
\description{
Tools to find edges and detect lines.
}
\details{
For Hough-line detection, the geometry format is \code{{W}x{H}+{threshold}}
defining the size and threshold of the filter used to find 'peaks' in
the intermediate search image. For canny edge detection the format is
\code{{radius}x{sigma}+{lower\%}+{upper\%}}. More details and examples are
available at the \href{https://www.imagemagick.org/Usage/transform}{imagemagick website}.
}
\examples{
shape <- image_read('http://www.imagemagick.org/Usage/transform/shape_rectangle.gif')
rectangle <- image_canny(shape)
rectangle \%>\% image_hough_draw('5x5+20')
rectangle \%>\% image_hough_txt(format = 'svg') \%>\% cat()
}
