% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{geometry}
\alias{geometry}
\alias{coordinate}
\title{Geometry Helpers}
\usage{
coordinate(x, y)

geometry(width, height, x_off = 0, y_off = 0, percent = FALSE,
  preserve_aspect = TRUE)
}
\arguments{
\item{x}{left offset in pixels}

\item{y}{top offset in pixels}

\item{width}{in pixels}

\item{height}{in pixels}

\item{x_off}{offset in pixels on x axis}

\item{y_off}{offset in pixels on y axis}

\item{percent}{interpret width and height as a percentage of the current size.}

\item{preserve_aspect}{if FALSE, resize to width and height exactly, loosing original aspect ratio.}
}
\description{
ImageMagick uses a special geometry syntax to specify coordinates and shapes.
You can simply specify these as strings but you can also use these functions
to construct them.
}
\details{
See \href{http://www.imagemagick.org/Magick++/Geometry.html}{ImageMagick Manual}
for syntax specification.
}
\examples{
coordinate(50, 80)
geometry(100, 200)
geometry(100, 200, 10, 10)
}
\references{
\href{http://www.imagemagick.org/Magick++/Geometry.html}{ImageMagick Manual}
}
