% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{image-attributes}
\alias{image-attributes}
\alias{image_info}
\title{Image Attributes}
\usage{
image_info(image)
}
\arguments{
\item{image}{object returned by \code{image_read}}
}
\description{
Functions to get or set Image attributes.
}
\seealso{
Other image: \code{\link{image-manipulate}},
  \code{\link{image-read}}
}

