% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image-manipulate.R
\name{image-manipulate}
\alias{image-manipulate}
\alias{image_background}
\alias{image_blur}
\alias{image_border}
\alias{image_crop}
\alias{image_delay}
\alias{image_format}
\alias{image_matte}
\alias{image_noise}
\alias{image_pen}
\alias{image_scale}
\alias{image_trim}
\title{Manipulating Images}
\usage{
image_format(image, format)

image_delay(image, delay)

image_trim(image)

image_background(image, color)

image_matte(image, matte = TRUE, color = "")

image_pen(image, color)

image_crop(image, geometry = "")

image_scale(image, geometry = "")

image_border(image, color = "", geometry = "")

image_noise(image, noisetype = 2)

image_blur(image, radius = 1, sigma = 0.5)
}
\arguments{
\item{image}{object returned by \code{image_read}}

\item{format}{output format such as \code{png}, \code{jpeg}, \code{gif} or \code{pdf}.}

\item{delay}{time in 1/100ths of a second (0 to 65535) which must expire before displaying
the next image in an animated sequence.}

\item{color}{a valid \href{https://www.imagemagick.org/Magick++/Color.html}{color string}}

\item{matte}{if the image has transparency. If set True, store matte channel
if the image has one otherwise create an opaque one.}

\item{geometry}{a string with \href{https://www.imagemagick.org/Magick++/Geometry.html}{geometry syntax}
for example \code{"10x10+5-5"}.}

\item{noisetype}{integer betwee 0 and 5 with
\href{https://www.imagemagick.org/Magick++/Enumerations.html#NoiseType}{noisetype}.}

\item{radius}{the radius of the Gaussian, in pixels, not counting the center pixel.}

\item{sigma}{the standard deviation of the Laplacian, in pixels.}
}
\description{
Vectorized functions for manipulating images.
}
\examples{
logo <- image_read(system.file("Rlogo.png", package = "magick"))
image_noise(logo)
logo <- image_read(system.file("Rlogo.png", package = "magick"))
image_blur(logo, 10, 10)
}
\seealso{
Other image: \code{\link{image-stl}}
}

