% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{image-read}
\alias{image-read}
\alias{image_append}
\alias{image_average}
\alias{image_browse}
\alias{image_coalesce}
\alias{image_display}
\alias{image_fft}
\alias{image_flatten}
\alias{image_join}
\alias{image_map}
\alias{image_montage}
\alias{image_morph}
\alias{image_mosaic}
\alias{image_read}
\alias{image_write}
\title{Image Manipulation}
\usage{
image_read(path)

image_write(image, path = NULL, format = NULL)

image_display(image, animate = TRUE)

image_browse(image, browser = getOption("browser"))

image_append(image, stack = FALSE)

image_average(image)

image_coalesce(image)

image_flatten(image)

image_fft(image)

image_map(image, map_image, dither = FALSE)

image_montage(image)

image_morph(image, frames)

image_mosaic(image)

image_join(imagelist)
}
\arguments{
\item{path}{file path, URL, or raw vector with image data}

\item{image}{object returned by \code{image_read}}

\item{format}{output format such as \code{png}, \code{jpeg}, \code{gif} or \code{pdf}.}

\item{animate}{support animations in the X11 display}

\item{browser}{argument passed to \link[utils:browseURL]{browseURL}}

\item{stack}{place images top-to-bottom (TRUE) or left-to-right (FALSE)}

\item{map_image}{reference image to map colors from}

\item{dither}{set TRUE to enable dithering}

\item{frames}{number of frames to use in output animation}

\item{imagelist}{list with images to be combined into a single image vector}
}
\description{
Read, write and edit images. All image functions are vectorized, meaning
they can operate either on a single frame or a series of frames (e.g. to
create a collage, video, or animation).
}
\examples{
# Download image from the web
frink <- image_read("https://jeroenooms.github.io/images/frink.png")
frink2 <- image_crop(frink)
image_write(frink2, "output.png")
# Create thumbnails from GIF
banana <- image_read(system.file("banana.gif", package = "magick"))
length(banana)
image_average(banana)
image_flatten(banana)
image_append(banana)
image_append(banana, stack = TRUE)

# Append images together
image_append(image_scale(c(image_append(banana[c(1,3)], stack = TRUE), frink)))
# Combine with another image
logo <- image_read(system.file("Rlogo.png", package = "magick"))
oldlogo <- image_read(system.file("Rlogo-old.png", package = "magick"))

# Create morphing animation
both <- image_scale(c(oldlogo, logo), "400x400")
image_average(image_crop(c(oldlogo, logo)))
animation <- image_morph(both, 10)
image_format(animation, "gif")
# Break down and combine frames
front <- image_scale(banana, "300")
background <- image_scale(logo, "400")
frames <- lapply(as.list(front), function(x) image_flatten(c(background, x)))
image_format(image_join(frames), "gif")
}
\references{
Magick++ Image STL: \url{https://www.imagemagick.org/Magick++/STL.html}
}
\seealso{
Other image: \code{\link{image-attributes}},
  \code{\link{image-manipulate}}
}

