% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_capture.R
\name{image_capture}
\alias{image_capture}
\title{Capture the contents of the current graphics device as an image}
\usage{
image_capture(device = dev.cur())
}
\arguments{
\item{device}{The graphics device you wish to capture.  The graphics device
must be a bitmap device that supports raster capture, such as Quartz, X11,
or devices in the Cairo package. Cairo(type="raster") is suggested for
non-interactive applications where you wish to avoid any writing to disk.}
}
\value{
A magick image pointer
}
\description{
Capture the contents of the current graphics device as an image
}
\examples{
library(Cairo)
Cairo(type="raster")
plot(density(rnorm(100)))
densimage <- image_capture()
dev.off()
image_flip(densimage)

}
