% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{image_convert}
\alias{image_convert}
\title{Convert Image}
\usage{
image_convert(image, to = "jpg")
}
\arguments{
\item{image}{file path or raw vector to input image}

\item{to}{output format, e.g. \code{jpeg}, \code{png}, etc}
}
\description{
Converts image to another format.
}
\examples{
png_file <- tempfile(fileext = ".png")
png(png_file)
plot(cars)
dev.off()
image_convert(png_file, "jpg")
image_convert(png_file, "gif")
image_convert(png_file, "pdf")
image_convert(png_file, "tiff")
}

