% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram.R
\name{image_histogram}
\alias{image_histogram}
\title{image histogram}
\usage{
image_histogram(image, height = 100)
}
\arguments{
\item{image}{an image}

\item{height}{the height of each histogram in pixels}
}
\value{
the result histograms
}
\description{
image histogram
}
\examples{
cute_cat <- image_read("http://i.telegraph.co.uk/multimedia/archive/02830/cat_2830677b.jpg")
image_histogram(cute_cat)
}

