% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treshold.R
\name{image_lat}
\alias{image_lat}
\title{Local Adaptive Thresholding}
\usage{
image_lat(image, geometry = "5x5+2\%")
}
\arguments{
\item{image}{magick image object returned by \code{\link[=image_read]{image_read()}} or \code{\link[=image_graph]{image_graph()}}}

\item{geometry}{pixel window plus offset for LAT algorithm}
}
\description{
Width x height define the size of the pixel neighborhood and offset
is a constant to subtract from pixel neighborhood mean.
}
