% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocr.R
\name{image_ocr}
\alias{image_ocr}
\title{OCR}
\usage{
image_ocr(image, ...)
}
\arguments{
\item{image}{object returned by \code{image_read}}

\item{...}{additional parameters passed to \code{tesseract::tesseract()}}
}
\description{
Extract text from an image. Requires the 'tesseract' package.
}
\examples{
if(require("tesseract")){
img <- image_read("http://jeroen.github.io/images/testocr.png")
image_ocr(img)
}
}
