% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold.R
\name{image_threshold}
\alias{image_threshold}
\title{Image thresholding}
\usage{
image_threshold(image, type = c("lat", "black", "white"),
  geometry = "5x5+2\%", threshold = "50\%")
}
\arguments{
\item{image}{magick image object returned by \code{\link[=image_read]{image_read()}} or \code{\link[=image_graph]{image_graph()}}}

\item{type}{type of thresholding, either one of lat, black or white (see details below)}

\item{geometry}{pixel window plus offset for LAT algorithm}

\item{threshold}{pixel intensity threshold percentage for black or white thresholding}
}
\description{
Thresholding an image can be used for simple and straightforward image segmentation.
The function \code{image_threshold} allows to do black thresholding, white thresholding
as well as local adaptive thresholding.
}
\details{
\itemize{
\item \code{type = black}: Forces all pixels below the threshold into black while leaving all pixels
at or above the threshold unchanged
\item \code{type = white}: Forces all pixels above the threshold into white while leaving all pixels
at or below the threshold unchanged
\item \code{type = lat}: Local Adaptive Thresholding. Looks in a box (width x height) around the
pixel neighborhood if the pixel value is bigger than the average minus an offset
}
}
\examples{
logo <- image_convert(logo, colorspace = "Gray")
image_threshold(logo, type = "lat", geometry = "5x5+10")
image_threshold(logo, type = "black", threshold = "50\%")
image_threshold(logo, type = "white", threshold = "50\%")

image_threshold(logo, type = "white", threshold = "50\%") \%>\%
  image_threshold(logo, type = "black", threshold = "50\%")
}
