% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gifski.R
\name{image_write_gif}
\alias{image_write_gif}
\title{Write to GIF}
\usage{
image_write_gif(image, path = NULL, width = NULL, height = NULL, ...)
}
\arguments{
\item{image}{magick image object returned by \code{\link[=image_read]{image_read()}} or \code{\link[=image_graph]{image_graph()}}}

\item{path}{filename of the output gif. This is also the return value of the function.}

\item{width}{output width in pixels. Defaults to first frame of input image.}

\item{height}{output height in pixels.}

\item{...}{extra parameters passed to \link[gifski:gifski]{gifski}}
}
\description{
Optimized implementation to generate animated gif using \link[gifski:gifski]{gifski}.
}
\details{
Accomplishes the same thing as \link{image_animate} but with higher quality
output. Requires that the gifski package is installed.
}
