% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagelist.R
\name{imagelist_read}
\alias{imagelist_read}
\alias{imagelist_write}
\title{Image List}
\usage{
imagelist_read(imagelist)

imagelist_write(imagelist, format = "gif", delay = 100, path = NULL)
}
\arguments{
\item{imagelist}{vector of images. Each element is passed to \link{image_read}.}

\item{format}{output format which supports animations}

\item{delay}{time in 1/100ths of a second (0 to 65535) which must expire before displaying
the next image in an animated sequence.}

\item{path}{file to write output. Use \code{NULL} to return as raw vector.}
}
\description{
Create and manipulate sequences of images, for example to generate animation.
}
\examples{
images <- sapply(1:5, function(i){
  png(tmp <- tempfile(), width = 800, height = 600)
  par(ask = FALSE)
  plot(rnorm(100))
  dev.off()
  return(tmp)
})
animation <- imagelist_read(images)
imagelist_write(animation, path = "out.gif")
}

