% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/device.R
\name{magick}
\alias{magick}
\title{Magick Graphics Device}
\usage{
magick(width = 1280, height = 720, bg = "transparent", pointsize = 12)
}
\arguments{
\item{width}{in pixels}

\item{height}{in pixels}

\item{bg}{background color}

\item{pointsize}{size of fonts}
}
\description{
Graphics device that produces a Magick image type.
}
