% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\name{frame_read}
\alias{frame_info}
\alias{frame_read}
\alias{frame_write}
\title{Magick Frames}
\usage{
frame_read(frame)

frame_write(frame, format, path = NULL)

frame_info(frame)
}
\arguments{
\item{frame}{magick-frame object, file path or raw vector with frame data}

\item{format}{output format, e.g. \code{jpeg}, \code{png}, \code{pdf} or \code{gif}.}

\item{path}{file to write output. Use \code{NULL} to return as raw vector.}
}
\description{
A frame is a single slice of an frame.
}
\examples{
png_file <- tempfile(fileext = ".png")
png(png_file)
plot(cars)
dev.off()
frame <- frame_read(png_file)
frame_info(frame)
frame_write(frame, "jpg", "output.jpg")
frame_write(frame, "gif", "output.gif")
frame_write(frame, "pdf", "output.pdf")
frame_write(frame, "tiff", "output.tiff")
}
\references{
frameMagick documentation: \url{https://www.framemagick.org/Magick++/image++.html}
}

