% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image.R
\name{image_read}
\alias{image_read}
\alias{image_write}
\title{Magick Images}
\usage{
image_read(image)

image_write(image, format, path = NULL)
}
\arguments{
\item{image}{magick-image object, file path or raw vector with image data}

\item{format}{output format, e.g. \code{jpeg}, \code{png}, \code{pdf} or \code{gif}.}

\item{path}{file to write output. Use \code{NULL} to return as raw vector.}
}
\description{
Converts image to another format.
}
\examples{
png_file <- tempfile(fileext = ".png")
png(png_file)
plot(cars)
dev.off()
image_write(png_file, "jpg")
image_write(png_file, "gif")
image_write(png_file, "pdf")
image_write(png_file, "tiff")
}

