% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocr.R
\name{ocr}
\alias{ocr}
\alias{image_ocr}
\title{Text OCR}
\usage{
image_ocr(image, language = "eng", ...)
}
\arguments{
\item{image}{object returned by \code{image_read}}

\item{language}{passed to \link[tesseract:tesseract]{tesseract}. To install additional languages see
instructions in \link[tesseract:tesseract_download]{tesseract_download}.}

\item{...}{additional parameters passed to \link[tesseract:tesseract]{tesseract}}
}
\description{
Extract text from an image using the \link[tesseract:tesseract]{tesseract} package. To use this function
you need to tesseract first using: \code{install.packages("tesseract")}.
}
\details{
Best results are obtained if you set the correct language in \link[tesseract:tesseract]{tesseract}.
To install additional languages see instructions in \link[tesseract:tesseract_download]{tesseract_download}.
}
\examples{
if(require("tesseract")){
img <- image_read("http://jeroen.github.io/images/testocr.png")
image_ocr(img)
}
}
