% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{options}
\alias{options}
\alias{option_values}
\alias{option_types}
\alias{option_dump}
\title{Magick Options}
\usage{
option_values(type)

option_types()

option_dump()
}
\arguments{
\item{type}{a string with any of the \code{option_types()}}
}
\description{
List option types and option values supported in your version of ImageMagick. For
details about each options see
\href{https://www.imagemagick.org/Magick++/Enumerations.html}{ImageMagick Enumerations}.
}
\examples{
# Check possible values:
option_values("Compose")
option_values("Morphology")

}
\references{
ImageMagick Manual: \href{https://www.imagemagick.org/Magick++/Enumerations.html}{Enumerations}
}
