% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resize.R
\name{resize}
\alias{resize}
\alias{image_trim}
\alias{image_chop}
\alias{image_rotate}
\alias{image_resize}
\alias{image_scale}
\alias{image_sample}
\alias{image_crop}
\alias{image_flip}
\alias{image_flop}
\alias{image_deskew}
\alias{image_page}
\title{Resize, Cut, Crop, Rotate}
\usage{
image_trim(image)

image_chop(image, geometry)

image_rotate(image, degrees)

image_resize(image, geometry = NULL, filter = NULL)

image_scale(image, geometry = NULL)

image_sample(image, geometry = NULL)

image_crop(image, geometry = NULL)

image_flip(image)

image_flop(image)

image_deskew(image, treshold = 40)

image_page(image, pagesize = NULL, density = NULL)
}
\arguments{
\item{image}{object returned by \code{image_read}}

\item{geometry}{a string with \href{https://www.imagemagick.org/Magick++/Geometry.html}{geometry syntax}
for example \code{"10x10+5-5"}}

\item{degrees}{value between 0 and 360 for how many degrees to rotate}

\item{filter}{string with a \href{https://www.imagemagick.org/Magick++/Enumerations.html#FilterTypes}{filtertype}.}

\item{treshold}{straightens an image. A threshold of 40 works for most images.}

\item{pagesize}{geometry string with preferred size and location of an image canvas}

\item{density}{geometry string with vertical and horizontal resolution in pixels of
the image. Specifies an image density when decoding a Postscript or PDF.}
}
\description{
Basic transformation and cutting operations.
}
\details{
The most powerful resize function is \code{image_resize} which allows for setting
a custom resize filter. Functions \code{image_scale} and \code{image_sample} are
similar to \code{image_resize(img, filter = "point")}.

For resize operations it holds that if no \code{geometry} is specified, all frames
are rescaled to match the top frame.
}
\examples{
logo <- image_read("logo:")
logo <- image_scale(logo, "400")
image_trim(logo)
image_chop(logo, "100x20")
image_rotate(logo, 45)
# Small image
rose <- image_convert(image_read("rose:"), "png")

# Resize to 400 width or height:
image_resize(rose, "400x")
image_resize(rose, "x400")

# Resize keeping ratio
image_resize(rose, "400x400")

# Resize, force size losing ratio
image_resize(rose, "400x400!")

# Different filters
image_resize(rose, "400x", filter = "Triangle")
image_resize(rose, "400x", filter = "Point")
# simple pixel resize
image_scale(rose, "400x")
image_sample(rose, "400x")
image_crop(logo, "400x400+200+200")
image_flip(logo)
image_flop(logo)
}
\seealso{
Other image: \code{\link{animation}},
  \code{\link{attributes}}, \code{\link{contrast}},
  \code{\link{editing}}, \code{\link{transformations}}
}
