% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{image_compare}
\alias{image_compare}
\title{Image Analysis}
\usage{
image_compare(image, reference_image, metric = "")
}
\arguments{
\item{image}{object returned by \code{image_read}}

\item{reference_image}{another image to compare to}

\item{metric}{string with a
\href{http://www.imagemagick.org/script/command-line-options.php#metric}{metric type}}
}
\description{
Functions for performing calculation and statistics on images.
}
\examples{
logo <- image_read("logo:")
logo2 <- image_blur(logo, 3)
logo3 <- image_oilpaint(logo, 3)
if(magick_config()$version >= "6.8.7"){
  image_compare(logo, logo2, metric = "phash")
  image_compare(logo, logo3, metric = "phash")
}
}
\seealso{
Other image: \code{\link{animation}},
  \code{\link{attributes}}, \code{\link{color}},
  \code{\link{composite}}, \code{\link{device}},
  \code{\link{editing}}, \code{\link{effects}},
  \code{\link{painting}}, \code{\link{resize}}
}
