% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wizard.R
\docType{data}
\name{wizard}
\alias{wizard}
\alias{logo}
\alias{rose}
\alias{granite}
\title{Example Images}
\format{
An object of class \code{magick-image} of length 1.
}
\usage{
logo
}
\description{
Example images included with ImageMagick:
}
\details{
\itemize{
\item \code{logo}: ImageMagick Logo, 640x480
\item \code{wizard}: ImageMagick Wizard, 480x640
\item \code{rose} : Picture of a rose, 70x46
\item \code{granite} : Granite texture pattern, 128x128
}
}
\keyword{datasets}
