% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{attributes}
\alias{attributes}
\alias{image_comment}
\title{Image Attributes}
\usage{
image_comment(image, comment = NULL)
}
\arguments{
\item{image}{object returned by \code{image_read}}

\item{comment}{string to set an image comment}
}
\description{
Attributes are properties of the image that might be present on some images
and might affect image manipualation methods. Each attribute can be get and
set with the same function.
}
\seealso{
Other image: \code{\link{animation}}, \code{\link{color}},
  \code{\link{composite}}, \code{\link{device}},
  \code{\link{editing}}, \code{\link{effects}},
  \code{\link{image_compare}}, \code{\link{painting}},
  \code{\link{resize}}
}
