% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{attributes}
\alias{attributes}
\alias{image_comment}
\alias{image_info}
\title{Image Attributes}
\usage{
image_comment(image, comment = NULL)

image_info(image)
}
\arguments{
\item{image}{magick image object returned by \code{\link[=image_read]{image_read()}} or \code{\link[=image_graph]{image_graph()}}}

\item{comment}{string to set an image comment}
}
\description{
Attributes are properties of the image that might be present on some images
and might affect image manipulation methods.
}
\details{
Each attribute can be get and set with the same function. The \code{\link[=image_info]{image_info()}}
function returns a data frame with some commonly used attributes.
}
\seealso{
Other image: \code{\link{_index_}}, \code{\link{analysis}},
  \code{\link{animation}}, \code{\link{color}},
  \code{\link{composite}}, \code{\link{device}},
  \code{\link{edges}}, \code{\link{editing}},
  \code{\link{effects}}, \code{\link{fx}},
  \code{\link{geometry}}, \code{\link{morphology}},
  \code{\link{ocr}}, \code{\link{options}},
  \code{\link{painting}}, \code{\link{segmentation}},
  \code{\link{transform}}, \code{\link{video}}
}
\concept{image}
