% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast.R
\name{contrast}
\alias{contrast}
\alias{image_contrast}
\alias{image_normalize}
\alias{image_modulate}
\alias{image_colorize}
\alias{image_enhance}
\alias{image_equalize}
\alias{image_median}
\alias{image_quantize}
\alias{image_transparent}
\title{Image Contrast and Colors}
\usage{
image_contrast(image, sharpen = 1)

image_normalize(image)

image_modulate(image, brightness = 100, saturation = 100, hue = 100)

image_colorize(image, opacity, color)

image_enhance(image)

image_equalize(image)

image_median(image, radius = 1)

image_quantize(image, max = 256, colorspace = NULL, dither = NULL,
  treedepth = NULL)

image_transparent(image, color, fuzz = 0)
}
\arguments{
\item{image}{object returned by \code{image_read}}

\item{sharpen}{enhance intensity differences in image}

\item{brightness}{modulation of brightness as percentage of the current value (100 for no change)}

\item{saturation}{modulation of saturation as percentage of the current value (100 for no change)}

\item{hue}{modulation of hue is an absolute rotation of -180 degrees to +180 degrees from the
current position corresponding to an argument range of 0 to 200 (100 for no change)}

\item{opacity}{percentage of opacity used for coloring}

\item{color}{a valid \href{https://www.imagemagick.org/Magick++/Color.html}{color string} such as
\code{"navyblue"} or \code{"#000080"}}

\item{radius}{radius, in pixels, for various transformations}

\item{max}{preferred number of colors in the image. The actual number of colors in the image may
be less than your request, but never more.}

\item{colorspace}{string with a magick \href{https://www.imagemagick.org/Magick++/Enumerations.html#ColorspaceType}{ColorspaceType}
for example 'grey' or 'rgb' or 'cmyk'}

\item{dither}{apply Floyd/Steinberg error diffusion to the image: averages intensities of sseveral
neighboring pixels}

\item{treedepth}{depth of the quantization color classification tree. Values of 0 or 1 allow
selection of the optimal tree depth for the color reduction algorithm. Values between 2 and 8
may be used to manually adjust the tree depth.}

\item{fuzz}{Colors within this distance are considered equal.
Use this option to match colors that are close to the target color in RGB space.
I think max distance (from #000000 to #FFFFFF) is 256^3.}
}
\description{
Functions to adjust constrast and colors.
}
\examples{
logo <- image_read("logo:")
image_contrast(logo)

# Convert to black-white
image_convert(logo, type = 'grayscale')
image_normalize(logo)
image_modulate(logo, brightness = 200)
image_modulate(logo, saturation = 150)
image_modulate(logo, hue = 200)
image_colorize(logo, 50, "red")
image_enhance(logo)
image_equalize(logo)
image_median(logo)
# Quantize into 10 colors, using various spaces
image_quantize(logo, max = 10, colorspace = 'gray')
image_quantize(logo, max = 10, colorspace = 'rgb')
image_quantize(logo, max = 10, colorspace = 'cmyk')
}
\seealso{
Other image: \code{\link{attributes}},
  \code{\link{editing}}, \code{\link{resize}},
  \code{\link{transformations}}
}
