% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast.R
\name{contrast}
\alias{contrast}
\alias{image_contrast}
\alias{image_normalize}
\alias{image_modulate}
\alias{image_colorize}
\alias{image_enhance}
\alias{image_equalize}
\alias{image_median}
\title{Image Contrast and Colors}
\usage{
image_contrast(image, sharpen = 1)

image_normalize(image)

image_modulate(image, brightness = 100, saturation = 100, hue = 100)

image_colorize(image, opacity, color)

image_enhance(image)

image_equalize(image)

image_median(image, radius = 1)
}
\arguments{
\item{image}{object returned by \code{image_read}}

\item{sharpen}{enhance intensity differences in image}

\item{brightness}{modulation of brightness as percentage of the current value (100 for no change)}

\item{saturation}{modulation of saturation as percentage of the current value (100 for no change)}

\item{hue}{modulation of hue is an absolute rotation of -180 degrees to +180 degrees from the
current position corresponding to an argument range of 0 to 200 (100 for no change)}

\item{opacity}{percentage of opacity used for coloring}

\item{color}{a valid \href{https://www.imagemagick.org/Magick++/Color.html}{color string} such as
\code{"navyblue"} or \code{"#000080"}}

\item{radius}{radius, in pixels, for various transformations}
}
\description{
Functions to adjust constrast and colors.
}
\examples{
logo <- image_read("logo:")
image_contrast(logo)

# Convert to black-white
image_convert(logo, type = 'grayscale')
image_normalize(logo)
image_modulate(logo, brightness = 200)
image_modulate(logo, saturation = 150)
image_modulate(logo, hue = 200)
image_colorize(logo, 50, "red")
image_enhance(logo)
image_equalize(logo)
image_median(logo)
}
\seealso{
Other image: \code{\link{editing}},
  \code{\link{transformations}}
}
