% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/device.R
\name{device}
\alias{device}
\alias{magick_device}
\title{Magick Graphics Device}
\usage{
magick_device(width = 800, height = 600, bg = "transparent",
  pointsize = 12, res = 72, clip = TRUE)
}
\arguments{
\item{width}{in pixels}

\item{height}{in pixels}

\item{bg}{background color}

\item{pointsize}{size of fonts}

\item{res}{resolution in pixels}

\item{clip}{enable clipping in the device. Because clippig can slow things down
a lot, you can disable it if you don't need it.}
}
\description{
Graphics device that produces a Magick image type. Works like other graphics
devices but instead of writing to a file, the drawings get written to a magick
image object which can then be treated like any other image object.
}
\details{
The device is a relatively recent feature of the package. It should support all
operations but there might still be small inaccuracies. Also it is currently a
bit slow, probably because it is rendering too frequently. Perhaps the next
version will use lazy rendering.
}
\examples{
img <- magick_device()
example(rect)
dev.off()
print(img)
}
