% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/device.R
\name{device}
\alias{device}
\alias{magick_device}
\title{Magick Graphics Device}
\usage{
magick_device(width = 1280, height = 720, bg = "transparent",
  pointsize = 12, clipping = FALSE)
}
\arguments{
\item{width}{in pixels}

\item{height}{in pixels}

\item{bg}{background color}

\item{pointsize}{size of fonts}

\item{clipping}{enable clipping support. This id currently disabled by default because
it slows things down a lot}
}
\description{
Graphics device that produces a Magick image type.
}
