% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image-manipulate.R
\name{image-manipulate}
\alias{image-manipulate}
\alias{image_border}
\alias{image_delay}
\alias{image_format}
\alias{image_trim}
\title{Manipulating Images}
\usage{
image_format(image, format)

image_delay(image, delay)

image_trim(image)

image_border(image, color = "", geometry = "")
}
\arguments{
\item{image}{object returned by \code{image_read}}

\item{format}{output format such as \code{png}, \code{jpeg}, \code{gif} or \code{pdf}.}

\item{delay}{time in 1/100ths of a second (0 to 65535) which must expire before displaying
the next image in an animated sequence.}

\item{color}{a valid \href{https://www.imagemagick.org/Magick++/Color.html}{color string}}

\item{geometry}{a string with \href{https://www.imagemagick.org/Magick++/Geometry.html}{geometry syntax}
for example \code{"10x10+5-5"}.}
}
\description{
Vectorized functions for manipulating images.
}
\seealso{
Other image: \code{\link{image-stl}}
}

