% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image-stl.R
\name{image-stl}
\alias{image-stl}
\alias{image_append}
\alias{image_read}
\alias{image_write}
\title{Image Manipulation}
\usage{
image_read(path)

image_append(image, stack = FALSE)

image_write(image, path = NULL)
}
\arguments{
\item{path}{file path, URL, or raw vector with image data}

\item{image}{object returned by \code{image_read}}

\item{stack}{if false, rectangular image frames are stacked left-to-right otherwise top-to-bottom.}
}
\description{
Read, write and edit images. All image functions are vectorized, meaning
they can operate either on a single frame or a series of frames (e.g. to
create a collage, video, or animation).
}
\examples{
library(magrittr)
img <- image_read("https://jeroenooms.github.io/images/frink.png")
print(img)

# trim whitespace and add a border
img \%>\% image_trim() \%>\% image_border("blue", "6x6+0+0")

# trim off the border again
img \%>\% image_trim()
}
\references{
Magick++ Image STL: \url{https://www.imagemagick.org/Magick++/STL.html}
}
\seealso{
Other image: \code{\link{image-manipulate}}
}

