% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image.R
\name{image_read}
\alias{image}
\alias{image_border}
\alias{image_delay}
\alias{image_format}
\alias{image_read}
\alias{image_trim}
\alias{image_write}
\alias{imagemagick}
\alias{magick}
\title{Image Manipulation}
\usage{
image_read(path)

image_format(image, format)

image_delay(image, delay)

image_trim(image)

image_border(image, color = "", geometry = "")

image_write(image, path = NULL)
}
\arguments{
\item{path}{file path, URL, or raw vector with image data}

\item{image}{object returned by \code{image_read}}

\item{format}{output format such as \code{png}, \code{jpeg}, \code{gif} or \code{pdf}.}

\item{delay}{time in 1/100ths of a second (0 to 65535) which must expire before displaying
the next image in an animated sequence.}

\item{color}{a valid \href{https://www.imagemagick.org/Magick++/Color.html}{color string}}

\item{geometry}{a string with \href{https://www.imagemagick.org/Magick++/Geometry.html}{geometry syntax}
for example \code{"10x10+5-5"}.}
}
\description{
Read, write and edit images. All image functions are vectorized, meaning
they can operate either on a single frame or a series of frames (e.g. to
work with video or animation).
}
\examples{
library(magrittr)
img <- image_read("https://jeroenooms.github.io/images/frink.png")
print(img)

# trim whitespace and add a border
img \%>\% image_trim() \%>\% image_border("blue", "6x6+0+0")

# trim off the border again
img \%>\% image_trim()
}
\references{
Magick++ Image STL: \url{https://www.imagemagick.org/Magick++/STL.html}
}

