% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treshold.R
\name{image_treshold}
\alias{image_treshold}
\title{Local Adaptive Thresholding}
\usage{
image_treshold(image, width, height, offset = 0)
}
\arguments{
\item{width}{size of the pixel neighborhood}

\item{height}{size of the pixel neighborhood}

\item{offset}{a constant to subtract from pixel neighborhood mean}
}
\description{
Width x height define the size of the pixel neighborhood and offset
is a constant to subtract from pixel neighborhood mean.
}
