% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image.R
\name{image_read}
\alias{image_info}
\alias{image_read}
\alias{image_write}
\title{Magick Images}
\usage{
image_read(image)

image_write(image, format, path = NULL)

image_info(image)
}
\arguments{
\item{image}{magick-image object, file path or raw vector with image data}

\item{format}{output format, e.g. \code{jpeg}, \code{png}, \code{pdf} or \code{gif}.}

\item{path}{file to write output. Use \code{NULL} to return as raw vector.}
}
\description{
Converts image to another format.
}
\examples{
png_file <- tempfile(fileext = ".png")
png(png_file)
plot(cars)
dev.off()
image <- image_read(png_file)
image_info(image)
image_write(image, "jpg", "output.jpg")
image_write(image, "gif", "output.gif")
image_write(image, "pdf", "output.pdf")
image_write(image, "tiff", "output.tiff")
}
\references{
ImageMagick documentation: \url{https://www.imagemagick.org/Magick++/Image++.html}
}

