% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{options}
\alias{options}
\alias{magick_options}
\alias{magick_fonts}
\alias{option_types}
\alias{filter_types}
\alias{metric_types}
\alias{dispose_types}
\alias{compose_types}
\alias{colorspace_types}
\alias{channel_types}
\alias{image_types}
\alias{kernel_types}
\alias{noise_types}
\alias{gravity_types}
\alias{orientation_types}
\alias{morphology_types}
\alias{style_types}
\alias{decoration_types}
\alias{compress_types}
\alias{distort_types}
\alias{dump_option_info}
\title{Magick Options}
\usage{
magick_options()

magick_fonts()

option_types()

filter_types()

metric_types()

dispose_types()

compose_types()

colorspace_types()

channel_types()

image_types()

kernel_types()

noise_types()

gravity_types()

orientation_types()

morphology_types()

style_types()

decoration_types()

compress_types()

distort_types()

dump_option_info(option = "font")
}
\arguments{
\item{option}{one of the \link{option_types}}
}
\description{
List option types and values supported in your version of ImageMagick. For
descriptions see
\href{https://www.imagemagick.org/Magick++/Enumerations.html}{ImageMagick Enumerations}.
}
\details{
The \link{dump_option_info} function is equivalent to calling \code{convert -list [option]} on
the command line. It does not return anything, it only makes ImageMagick print
stuff to the console, use only for debugging.
}
\references{
ImageMagick Manual: \href{https://www.imagemagick.org/Magick++/Enumerations.html}{Enumerations}
}
\seealso{
Other image: 
\code{\link{_index_}},
\code{\link{analysis}},
\code{\link{animation}},
\code{\link{attributes}()},
\code{\link{color}},
\code{\link{composite}},
\code{\link{defines}},
\code{\link{device}},
\code{\link{edges}},
\code{\link{editing}},
\code{\link{effects}()},
\code{\link{fx}},
\code{\link{geometry}},
\code{\link{morphology}},
\code{\link{ocr}},
\code{\link{painting}},
\code{\link{segmentation}},
\code{\link{transform}()},
\code{\link{video}}
}
\concept{image}
