% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{options}
\alias{options}
\alias{magick_options}
\alias{option_types}
\alias{filter_types}
\alias{metric_types}
\alias{dispose_types}
\alias{compose_types}
\alias{colorspace_types}
\alias{channel_types}
\alias{image_types}
\alias{kernel_types}
\alias{noise_types}
\alias{gravity_types}
\alias{orientation_types}
\title{Magick Options}
\usage{
magick_options()

option_types()

filter_types()

metric_types()

dispose_types()

compose_types()

colorspace_types()

channel_types()

image_types()

kernel_types()

noise_types()

gravity_types()

orientation_types()
}
\description{
List option types and values supported in your version of ImageMagick. For
descriptions see
\href{https://www.imagemagick.org/Magick++/Enumerations.html}{ImageMagick Enumerations}.
}
\references{
ImageMagick Manual: \href{https://www.imagemagick.org/Magick++/Enumerations.html}{Enumerations}
}
\seealso{
Other image: \code{\link{_index_}}, \code{\link{analysis}},
  \code{\link{animation}}, \code{\link{attributes}},
  \code{\link{color}}, \code{\link{composite}},
  \code{\link{device}}, \code{\link{editing}},
  \code{\link{effects}}, \code{\link{geometry}},
  \code{\link{ocr}}, \code{\link{painting}},
  \code{\link{transform}}
}
\concept{image}
